requirements.txt
numpy
pandas
streamlit
pytesseract
Pillow
google-generativeai
faiss-cpu
sentence-transformers
